Issue #Dockerfile: Added dockerfile and needed config changes
Made sure config can be loaded during runtime for Docker.

Changes to the frontend path referring to assets HAVE TO
use absolute paths, anything in public folder just needs a slash.
For instance `PROJECT_ROOT/apps/frontend/public/images/image.png`
can be accessed with `/images/image.png`, this is important
for docker image.

Added basic script in the root to build images, a bit simpler as
Dockerfiles have to be in the subproject folders but context needs
to be global (root of monorepo).